requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
cognee==0.1.39
numpy==1.26.4
pandas==2.2.3
pyyaml>=6.0.0
openai>=1.0.0